Update CelsiusFarneheitKelvin.c
Since new versions of the project are to be made continually, splitting each up into versions is important. Since this version is so new, you may notice it has a different title than later reworks. Depending on how drastic a change may be for a given update to the CFK.c program, versions may jump higher or lower amounts.